+++ source/commands/ci/_tests/runner.test.ts
@@ -93,3 +93,51 @@ it.skip("passes the dangerID from args into the executor config", async () => {
   const executor: Executor = mockRunDangerSubprocess.mock.calls[0][2]
   expect(executor.options.dangerID).toEqual("test-danger-run")
 })
+
+it("sets newComment to false by default", async () => {
+  await runRunner(defaultAppArgs as SharedCLI, { platform, source })
+
+  // Pull the executor out of the call to runDangerSubprocess
+  const executor: Executor = mockRunDangerSubprocess.mock.calls[0][2]
+  expect(executor.ciSource).toEqual(source)
+  expect(executor.platform).toEqual(platform)
+  expect(executor.options.newComment).toEqual(false)
+})
+
+it("passes newComment option if requested", async () => {
+  const customArgs = {
+    ...defaultAppArgs,
+    newComment: true,
+  } as SharedCLI
+  await runRunner(customArgs, { platform, source })
+
+  // Pull the executor out of the call to runDangerSubprocess
+  const executor: Executor = mockRunDangerSubprocess.mock.calls[0][2]
+  expect(executor.ciSource).toEqual(source)
+  expect(executor.platform).toEqual(platform)
+  expect(executor.options.newComment).toEqual(true)
+})
+
+it("sets removePreviousComments to false by default", async () => {
+  await runRunner(defaultAppArgs as SharedCLI, { platform, source })
+
+  // Pull the executor out of the call to runDangerSubprocess
+  const executor: Executor = mockRunDangerSubprocess.mock.calls[0][2]
+  expect(executor.ciSource).toEqual(source)
+  expect(executor.platform).toEqual(platform)
+  expect(executor.options.removePreviousComments).toEqual(false)
+})
+
+it("passes removePreviousComments option if requested", async () => {
+  const customArgs = {
+    ...defaultAppArgs,
+    removePreviousComments: true,
+  } as SharedCLI
+  await runRunner(customArgs, { platform, source })
+
+  // Pull the executor out of the call to runDangerSubprocess
+  const executor: Executor = mockRunDangerSubprocess.mock.calls[0][2]
+  expect(executor.ciSource).toEqual(source)
+  expect(executor.platform).toEqual(platform)
+  expect(executor.options.removePreviousComments).toEqual(true)
+})