+++ source/commands/ci/runner.ts
@@ -69,6 +69,8 @@ export const runRunner = async (app: SharedCLI, config?: Partial<RunnerConfig>) 
         failOnErrors: app.failOnErrors,
         noPublishCheck: !app.publishCheck,
         ignoreOutOfDiffComments: app.ignoreOutOfDiffComments,
+        newComment: app.newComment || false,
+        removePreviousComments: app.removePreviousComments || false,
       }
 
       const processName = (app.process && addSubprocessCallAguments(app.process.split(" "))) || undefined